third_party/blink/web_tests/wpt_internal/attribution-reporting/simple-verbose-debug-report.sub.https.html

<!doctype html>
<meta charset=utf-8>
<meta name=timeout content=long>
<script src="/common/get-host-info.sub.js"></script>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/helpers.js"></script>
<script>
attribution_reporting_promise_test(async t => {
  await registerAttributionSrc({
    trigger: {
      event_trigger_data: [{}],
      debug_reporting: true,
      debug_key: '456',
    },
    cookie: attributionDebugCookie,
  });

  const payload = await pollVerboseDebugReports();
  assert_equals(payload.reports.length, 1);
  const report = JSON.parse(payload.reports[0].body);
  assert_equals(report.length, 1);
  assert_equals(report[0].type, 'trigger-no-matching-source');
  assert_own_property(report[0], 'body');
  assert_equals(report[0].body.attribution_destination, 'https://{{host}}');
  assert_equals(report[0].body.trigger_debug_key, '456');
}, 'Verbose debug report is received.');
</script>
